National Initiative for Cybersecurity  Education  (N ICE )   Cybersecurity Workforce Framework
This publication describes the National Initiative for Cybersecurity Education (NICE)Cybersecurity Workforce Framework (NICE Framework), a reference structure that describesthe interdisciplinary nature of the cybersecurity work. It serves as a fundamental reference resource for describing and sharing information about cybersecurity work and the knowledge,skills, and abilities (KSAs) needed to complete tasks that can strengthen the cybersecurity posture of an organization. As a common, consistent lexicon that categorizes and describes cybersecurity work, the NICE Framework improves communication about how to identify,recruit, develop, and retain cybersecurity talent. The NICE Framework is a reference source from which organizations or sectors can develop additional publications or tools that meet their needs to define or provide guidance on different aspects of cybersecurity workforce development,planning, training, and education.